# docs/sap-openapi-source.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Secure Assessment Platform API",
    "description": "API for managing secure online assessments with proctoring features",
    "version": "1.0.0",
    "contact": {
      "name": "Secure Assessment Platform Support",
      "email": "support@secureassessment.example"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Development server"
    },
    {
      "url": "https://api.secureassessment.example/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/analytics/activity": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Report activity",
        "description": "Reports a user activity event",
        "operationId": "reportActivity",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityReportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Activity reported successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityReport"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/analytics/assessments/{id}/session": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Track assessment session",
        "description": "Reports a session-related event for an assessment",
        "operationId": "trackAssessmentSession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Assessment ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionTrackingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session event tracked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionTrackingResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/analytics/suspicious": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Log suspicious activity",
        "description": "Reports suspicious activity during an assessment",
        "operationId": "logSuspiciousActivity",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspiciousActivityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Suspicious activity logged successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspiciousActivityResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/dashboard/summary": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get dashboard summary",
        "description": "Returns summary statistics for the admin dashboard",
        "operationId": "getDashboardSummary",
        "responses": {
          "200": {
            "description": "Dashboard summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/dashboard/activity": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get activity timeline",
        "description": "Returns recent activity for the admin dashboard",
        "operationId": "getActivityTimeline",
        "responses": {
          "200": {
            "description": "Activity timeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityTimeline"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/system/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get system status",
        "description": "Returns the current system status",
        "operationId": "getSystemStatus",
        "responses": {
          "200": {
            "description": "System status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/attempt/grade": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Grade attempt by add feedback and final score.",
        "description": "Return result of update attempt",
        "operationId": "gradeAttempt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeAttempt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grade Attempt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeAttemptResult"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Not authenticated",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthenticationError"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Not authorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AuthorizationError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NotFoundError"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GradeAttempt": {
        "type": "object",
        "properties": {
          "feedback": {
            "type": "string",
            "example": "good work"
          },
          "score": {
            "type": "number",
            "format": "float",
            "example": 85
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "format": "int64",
                  "example": 1
                },
                "isCorrect": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        }
      },
...
      "ActivityReportRequest": {
        "type": "object",
        "required": [
          "userId",
          "action"
        ],
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 5
          },
          "action": {
            "type": "string",
            "example": "ASSESSMENT_VIEW"
          },
          "assessmentId": {
            "type": "integer",
            "format": "int64",
            "example": 3
          },
          "details": {
            "type": "string",
            "example": "Viewed assessment details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T15:42:18Z"
          }
        }
      },
      "ActivityReport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1256
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 5
          },
          "action": {
            "type": "string",
            "example": "ASSESSMENT_VIEW"
          },
          "assessmentId": {
            "type": "integer",
            "format": "int64",
            "example": 3
          },
          "details": {
            "type": "string",
            "example": "Viewed assessment details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T15:42:18Z"
          }
        }
      },
      "SessionTrackingRequest": {
        "type": "object",
        "required": [
          "userId",
          "action"
        ],
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 5
          },
          "action": {
            "type": "string",
            "enum": [
              "SESSION_START",
              "SESSION_END",
              "TAB_HIDDEN",
              "TAB_VISIBLE",
              "QUESTION_VIEWED",
              "QUESTION_ANSWERED"
            ],
            "example": "SESSION_START"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T14:30:00Z"
          },
          "userAgent": {
            "type": "string",
            "example": "Mozilla/5.0..."
          },
          "questionId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      },
      "SessionTrackingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 123
          },
          "assessmentId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 5
          },
          "action": {
            "type": "string",
            "enum": [
              "SESSION_START",
              "SESSION_END",
              "TAB_HIDDEN",
              "TAB_VISIBLE",
              "QUESTION_VIEWED",
              "QUESTION_ANSWERED"
            ],
            "example": "SESSION_START"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T14:30:00Z"
          }
        }
      },
      "SuspiciousActivityRequest": {
        "type": "object",
        "required": [
          "userId",
          "assessmentId",
          "type"
        ],
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 5
          },
          "assessmentId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "TAB_SWITCHING",
              "FACE_NOT_DETECTED",
              "MULTIPLE_FACES",
              "LOOKING_AWAY",
              "SUSPICIOUS_OBJECT",
              "VOICE_DETECTED"
            ],
            "example": "TAB_SWITCHING"
          },
          "details": {
            "type": "string",
            "example": "User switched tabs 3 times"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T14:35:22Z"
          },
          "userAgent": {
            "type": "string",
            "example": "Mozilla/5.0..."
          }
        }
      },
      "SuspiciousActivityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 45
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "example": 5
          },
          "assessmentId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "TAB_SWITCHING",
              "FACE_NOT_DETECTED",
              "MULTIPLE_FACES",
              "LOOKING_AWAY",
              "SUSPICIOUS_OBJECT",
              "VOICE_DETECTED"
            ],
            "example": "TAB_SWITCHING"
          },
          "details": {
            "type": "string",
            "example": "User switched tabs 3 times"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T14:35:22Z"
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "example": "MEDIUM"
          },
          "reviewed": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "DashboardSummary": {
        "type": "object",
        "properties": {
          "users": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 457
              },
              "active": {
                "type": "integer",
                "example": 428
              },
              "inactive": {
                "type": "integer",
                "example": 29
              },
              "newThisWeek": {
                "type": "integer",
                "example": 15
              }
            }
          },
          "assessments": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 42
              },
              "active": {
                "type": "integer",
                "example": 25
              },
              "draft": {
                "type": "integer",
                "example": 10
              },
              "expired": {
                "type": "integer",
                "example": 7
              },
              "newThisWeek": {
                "type": "integer",
                "example": 3
              }
            }
          },
          "activity": {
            "type": "object",
            "properties": {
              "assessmentAttempts": {
                "type": "object",
                "properties": {
                  "total": {
                    "type": "integer",
                    "example": 2785
                  },
                  "thisWeek": {
                    "type": "integer",
                    "example": 248
                  },
                  "passRate": {
                    "type": "number",
                    "format": "float",
                    "example": 73.5
                  }
                }
              },
              "usersOnline": {
                "type": "integer",
                "example": 24
              },
              "recentSuspiciousActivity": {
                "type": "integer",
                "example": 5
              }
            }
          }
        }
      },
...
      "ActivityTimeline": {
        "type": "object",
        "properties": {
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityTimelineEvent"
            }
          }
        }
      },
      "SystemStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded",
              "down"
            ],
            "example": "healthy"
          },
          "services": {
            "type": "object",
            "properties": {
              "database": {
                "type": "string",
                "enum": [
                  "operational",
                  "degraded",
                  "down"
                ],
                "example": "operational"
              },
              "storage": {
                "type": "string",
                "enum": [
                  "operational",
                  "degraded",
                  "down"
                ],
                "example": "operational"
              },
              "webcam": {
                "type": "string",
                "enum": [
                  "operational",
                  "degraded",
                  "down"
                ],
                "example": "operational"
              },
              "ai": {
                "type": "string",
                "enum": [
                  "operational",
                  "degraded",
                  "down"
                ],
                "example": "operational"
              }
            }
          },
          "statistics": {
            "type": "object",
            "properties": {
              "uptime": {
                "type": "string",
                "example": "15 days, 7 hours"
              },
              "activeConnections": {
                "type": "integer",
                "example": 56
              },
              "averageResponseTime": {
                "type": "integer",
                "example": 124
              },
              "cpuUsage": {
                "type": "number",
                "format": "float",
                "example": 35.2
              },
              "memoryUsage": {
                "type": "number",
                "format": "float",
                "example": 42.8
              }
            }
          },
          "lastChecked": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T15:54:32Z"
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T15:55:12Z"
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "Must be a valid email address"
                }
              }
            }
          }
        }
      },
      "AuthenticationError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "UNAUTHORIZED"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-09T15:55:12Z"
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/login"
          }
        }
      },
...
      "GradeAttemptResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "example": "Sucessfully graded attempt"
          }
        }
      }
    }
  }
}